Add service component FriendsListService
This service component contains some functions to access the database.
These functions will be used by the components that show the friends list.

The function allFriends obtains all user IDs of the friends of the user
with the specified user ID. This doesn't work at the moment, because the
sessionUser can't be determined.
There are also 3 empty functions:
 - endFriendship
 - acceptFriendRequest
 - declineFriendRequest