Refactor the whole InboxMailComponent
The component tried to load the mails and to add the received mails
by using jQuery. This resulted in very dirty and unmaintainable code
which also bypassed the element creation system of React. In fact the
mail items were appended to the page, but were totally decoupled from
React's mechanisms and therefore not able to interact with the
component's class or the provided event handlers.

The code base for the mail items has been extracted to its own
React component. The mail item component receives the data by passing
it as an attribute. The passed data contains an object containing
the message and the corresponding user. This way the mail item
component knows everything it needs to know for rendering or any
event handling. That said, the inbox component now creates only such
mail item objects and passes the needed data instead of creating
such objects itself. And because of that the mail item component is
a real React component it is able to do anythin React provides. This
was a problem with the elements created by jQuery.

The inbox uses also the provided promises inteface which leads to
write cleaner and simpler code that can be way better maintained.